requirements.txt
python-telegram-bot[job-queue]==20.7
yfinance==0.2.18
requests==2.31.0
httpx>=0.25.2
python-dotenv==1.0.0
aiohttp==3.9.1
cachetools==5.3.2
qrcode==8.0
pillow>=11.0.0,<12.0.0

